feat: Complete dashboard functionality
